requirements.txt
pandas==2.3.2
scipy==1.16.2
numpy==2.2.3
transformers==4.56.1
pillow==11.1.0
opencv-python==4.12.0
moviepy==1.0.3
scikit-learn==1.6.1
shapely==2.1.2
geopandas==1.1.1
statsmodels==0.14.5
matplotlib==3.10.6
seaborn==0.13.2
google-genai==1.41.0
pydantic==2.11.10